pyproject.toml
[project]
name = "s3vec-py"
version = "0.2.0"
description = "S3-compatible vector store (FastAPI, MinIO/S3 backend, hybrid ANN, FAISS + HNSW)"
readme = "README.md"
requires-python = ">=3.13.5"
authors = [{ name = "You" }]
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",
  "boto3>=1.34.0",
  "botocore>=1.34.0",
  "numpy>=1.24.0",
  "pyarrow>=15.0.0",
  "orjson>=3.9.15",
  "pydantic>=2.6.0",
  "python-multipart>=0.0.9",
  "faiss-cpu>=1.7.4",
  "hnswlib>=0.8.0",
  "python-dateutil>=2.8.2",
  "requests>=2.32.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
  "ruff>=0.4.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]
